refactor(config-helper): moves SchemaError to its own file
