+++ packages/config-helper/src/index.ts
@@ -6,3 +6,4 @@ export * from './lib/parm'
 export * from './lib/properties'
 export * from './lib/schema'
 export * from './lib/params'
+export * from './lib/schema.error'
