+++ packages/config-helper/src/lib/config-default-impl.ts
@@ -1,6 +1,7 @@
 import { Properties } from './properties'
-import { NotConvertable, NoValue, Schema } from './schema'
-import { Config, IllegalNullValue, SchemaError } from './config'
+import { NoValue, Schema } from './schema'
+import { NotConvertable, IllegalNullValue, SchemaError } from './schema.error'
+import { Config } from './config'
 import { isNormalizedSchemaObject, NormalizedConfigDefinition, NormalizeSchema } from './normalized-schema'
 import { ConfigOptions } from './config-options'
 import { normalizeSchema } from './create-config'
