+++ packages/config-helper/src/lib/params/string-transformer.spec.ts
@@ -1,6 +1,6 @@
 import { stringTransformer } from './string-transformer'
 import { err, ok } from 'neverthrow'
-import { NotConvertable } from '../schema'
+import { NotConvertable } from '../schema.error'
 
 describe('string-transformer', () => {
 	it.each([
